+++ scripts/ci/prepare_linux_integration_test_setup_only.sh
@@ -83,9 +83,6 @@ docker run -d -p 8086:8086 -e "DOCKER_INFLUXDB_INIT_MODE=setup" -e "DOCKER_INFLU
 # Start up influx (1 for influxql)
 docker run -d -p 8087:8086 -e "INFLUXDB_HTTP_AUTH_ENABLED=true" -e "INFLUXDB_ADMIN_USER=test" -e "INFLUXDB_ADMIN_PASSWORD=testtest" influxdb:1.7
 
-# Start up neo4j
-neo4j="$(docker run -d -p7474:7474 -p7687:7687 --env NEO4J_AUTH=neo4j/password neo4j)"
-
 # Start up mongodb and install mongosh (shell)
 docker run -d -e "MONGO_INITDB_ROOT_USERNAME=test" -e "MONGO_INITDB_DATABASE=test" -e "MONGO_INITDB_ROOT_PASSWORD=test" -p 27017:27017 mongo:5
 # curl -LO https://github.com/mongodb-js/mongosh/releases/download/v1.5.0/mongodb-mongosh_1.5.0_amd64.deb
@@ -127,9 +124,6 @@ retry 10 "curl -X PUT http://localhost:9200/test"
 for t in $(ls testdata/documents/*.json); do
     retry 10 "curl -X POST -H 'Content-Type: application/json' -d @$t http://localhost:9200/test/_doc"
 done
-
-# Configure neo4j
-docker exec "$neo4j" bin/cypher-shell -u neo4j -p password "CREATE (u:User { name : 'test' });"
 
 # Load Mongodb documents
 for t in $(ls testdata/documents/*.json); do
